requirements.txt
Django>=4.1.0,<5.0.0
djangorestframework>=3.14.0,<3.16.0
drf-yasg>=1.21.0,<2.0.0